add Decimal support

+++ src/schema/self_referential.rs
@@ -78,9 +78,11 @@ pub enum SchemaNode<'a> {
 	///
 	/// `scale` defaults to 0 and is an integer greater than or equal to 0 and `precision` is an
 	/// integer greater than 0.
+	///
+	/// https://avro.apache.org/docs/current/specification/#decimal
 	Decimal {
 		precision: usize,
-		scale: usize,
+		scale: u32,
 		inner: &'a SchemaNode<'a>,
 	},
 	/// A universally unique identifier, annotating a string.
